requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
referencing==0.36.2
rpds-py==0.25.0
sqlparse==0.5.3
typing_extensions==4.13.2
